requirements.txt
python-telegram-bot==22.1
beautifulsoup4==4.12.3
requests==2.31.0
feedparser==6.0.11
APScheduler==3.10.4 